--- bout_runners/metadata_recorder/__init__.py
@@ -1,1 +1,0 @@
-"""Package containing the metadata recorder modules."""
